New Folder/virtual-doctor-main/requirements.txt
google-generativeai
gradio
pdfplumber
pypdf
langchain
google-generativeai

# api_key='AIzaSyCTmK35I8Hr3HF-BHOwwlBkvZxVBZCceOc'